feat: Поменял список комнат на словарь

+++ models/server_rooms.py
@@ -10,30 +10,29 @@ from models.server_users import get_user
 from models.user import User
 from utils import generate_code
 
-server_rooms: List[Room] = []
+server_rooms_dict: dict[int, Room] = {}
 
 async def get_room_where_user(user_id) -> Room:
     logging.info(f'Get room containing USER_{user_id}')
-    room = [room for room in server_rooms if user_id in room.users or user_id in room.moderators or user_id in room.admins]
-    if room and len(room) > 0:
-        return room[0]
+    for r in server_rooms_dict.values():
+        if user_id in r.users or user_id in r.moderators or user_id in r.admins:
+            return r
     return None
 
 
 async def get_room(room_id) -> Room:
-    # TODO: Заменить список комнат на словарь по аналогии с пользователями
-    logging.info(f'Get room ROOM_{room_id}')
+    room_id = str(room_id)
     if room_id == '':
         return None
-    room = [room for room in server_rooms if room.room_id == room_id]
-    if len(room) > 0:
-        return room[0]
+    logging.info(f'Get room ROOM_{room_id}')
+    if room_id in server_rooms_dict:
+        return server_rooms_dict[room_id]
     else:
         logging.info(f'Try pooling from database')
         room = await try_get_room_from_db(room_id)
         if room:
             logging.info(f'Room found - {room.name}')
-            await add_room(room)
+            await add_room(room_id, room)
             return room
         else:
             logging.warning(f'Room was not found in database!')
@@ -72,7 +71,7 @@ def load_room_from_json(room_id, db_room) -> Room:
 
 
 def check_room_join_code(room_join_code):
-    for r in server_rooms:
+    for r in server_rooms_dict.values():
         if r.users_join_code == room_join_code:
             return True
     return False
@@ -90,31 +89,32 @@ async def create_room(user_id, room_name) -> Room:
 
     rooms_ref = db.reference('/rooms')
     room_ref = rooms_ref.push(room.to_dict())
-    room.set_room_id(room_ref.key)
+    room_id = room_ref.key
+    room.set_room_id(room_id)
 
     logging.info(f'Created: {room.to_dict()}')
-    await user.add_owned_room(room_ref.key)
-    await add_room(room)
+    await user.add_owned_room(room_id)
+    await add_room(room_id, room)
     return room
 
 
-async def add_room(room: Room):
-    logging.info(f'Add new room to cache | ROOM_{room.room_id}')
-    server_rooms.append(room)
+async def add_room(room_id, room: Room):
+    logging.info(f'Caching ROOM_{room.room_id}')
+    server_rooms_dict[room_id] = room
 
 
 async def remove_room(room_id, user_id):
     logging.info(f'Removing ROOM_{room_id} by USER_{user_id}')
-    rooms_to_remove = [room for room in server_rooms if room.room_id == room_id]
-    room_to_remove = rooms_to_remove[0]
-
+    room_to_remove = await get_room(room_id)
+    if room_to_remove is None:
+        return
     is_admin = await room_to_remove.is_user_admin(user_id)
     if is_admin:
         user: User = await get_user(user_id)
         await delete_room_event.fire(room_to_remove.get_users_list())
         await user.remove_owned_room(room_id)
         await room_to_remove.delete()
-        server_rooms.remove(room_to_remove)
+        del server_rooms_dict[room_id]
 
 
 async def get_room_by_join_code(join_code, user_role):
@@ -124,26 +124,26 @@ async def get_room_by_join_code(join_code, user_role):
         'moderator': 'mod_password'
     }
 
-    room = None
     if user_role == 'user':
-        room = [room for room in server_rooms if room.users_join_code == join_code]
+        for r in server_rooms_dict.values():
+            if r.users_join_code == join_code:
+                return r
     if user_role == 'moderator':
-        room = [room for room in server_rooms if room.moderators_join_code == join_code]
+        for r in server_rooms_dict.values():
+            if r.moderators_join_code == join_code:
+                return r
 
-    if room and len(room) > 0:
-        return room[0]
-    else:
-        logging.info(f'Try pooling from database')
-        rooms_ref = db.reference('/rooms')
-        db_room_data = rooms_ref.order_by_child(role_to_room_code[user_role]).equal_to(join_code).get()
-        rooms_list = list(db_room_data.items())
-        if len(rooms_list) < 1:
-            return None
-        room_key, room_data = rooms_list[0]
-        room = load_room_from_json(room_key, room_data)
-        if room:
-            await add_room(room)
-        return room
+    logging.info(f'Try pooling from database')
+    rooms_ref = db.reference('/rooms')
+    db_room_data = rooms_ref.order_by_child(role_to_room_code[user_role]).equal_to(join_code).get()
+    rooms_list = list(db_room_data.items())
+    if len(rooms_list) < 1:
+        return None
+    room_key, room_data = rooms_list[0]
+    room = load_room_from_json(room_key, room_data)
+    if room:
+        await add_room(room_key, room)
+    return room
 
 
 async def update_database_room_handler(room: Room):
